Removed the zodb magic, and all use of Connection.add() by refactoring the PersistentKeysDict to cope with persistent objects that don't yet have a _p_oid, and to cope with when the PersistentKeysDict does not have a _p_jar. Persistent objects that do not have a _p_oid are given one when __getstate__ is called on the PersistentKeysDict.
The change to the system as a whole is that clients of PersistentKeysDict
no longer need to ensure that it has a _p_jar before it is used.
Also, there is no need to do anything special for unit tests now.

Although all tests pass, there is a need to write some more tests testing
in detail PersistentKeysDict's interaction with the persistence machinery.