Made i18n in SchoolBell relationships restive view more idiomatic.
I'm not even sure if we need i18n in this particular case at all.  ReSTive
views are not for people to use directly anyway, so we might just as well make
the returned messages standard and more easily parseable.  And I highly doubt
that ReST requests would contain language information.