Step two of the timetable schema refactoring: ITimetableSchemaContainer now contains ITimetableSchemas and doesn't play games with cloneEmpty.  If you want to create a timetable from a schema, you have to explicitly call schema.createTimetable().
Expanded the ITimetableSchemaDay interface a bit to make it somewhat compatible
with ITimetableDay, so that timetable schemas can be passed to
ITimetableModel.getPeriodsForDay and rendered with ITimetable views.

Split ITimetableSchemaWrite off from ITimetableSchema and added securit
declarations.

Fixed code and tests to no longer create timetables where schemas are needed.

Added comparison methods to ITimetableSchema, to facilitate unit tests.

The next step (optional) is to refactor the implementation of Timetable, to
reduce duplication.