Site-wide calendar in everyones overlay.
We now have an event subscriber so that the site-wide calendar will by default
be in a new persons overlay portlet.  This can be de-activated by the user in
the overlay-selection view.

At the moment you need to enable viewCalendar permissions for the authenticated
group on the site-wide calendar to see the overlays in action.