generation script to make schooltool's application calendar public and set security policy to the 0.11 defaults.
I'm not sure if this is a good idea for back porting because the change in
security policy - from unset on containers, to denyAccess - may result in
unexpected behavior in existing sites.

It's probably the correct way for SchoolTool to set access from now on though.

Dragons:  The old evolve3.py is now evolve4.py.  This
shouldn't hurt people using release tarballs, but it may play havoc with
devlopers svn sandboxes, sorry about that.